+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add explicit `__all__` configuration
+- Support [PEP 604] unions through `types.UnionType`
+
+[PEP 604]: https://peps.python.org/pep-0604/
 
 ### Fixed